+++ framework/spring-boot-starter-mongodb-reactive/src/main/java/org/occurrent/springboot/mongo/reactor/OccurrentReactiveMongoAutoConfiguration.java
@@ -31,6 +31,7 @@ import org.occurrent.application.service.reactor.generic.GenericApplicationServi
 import org.occurrent.application.service.spring.reactor.SpringReactiveTransactionExecutor;
 import org.occurrent.dsl.dcb.reactor.DcbDomainEventQueries;
 import org.occurrent.dsl.dcb.reactor.DcbSubscriptions;
+import org.occurrent.dsl.projection.AppliedPositionStorage;
 import org.occurrent.dsl.query.reactor.DomainEventQueries;
 import org.occurrent.dsl.subscription.reactor.StreamSubscriptions;
 import org.occurrent.dsl.subscription.reactor.Subscriptions;
@@ -155,6 +156,16 @@ public class OccurrentReactiveMongoAutoConfiguration<E> {
     @Conditional(OnSubscriptionsNotDisabledCondition.class)
     public CheckpointStorage occurrentCheckpointStorage(ReactiveMongoOperations mongo, OccurrentProperties occurrentProperties) {
         return new ReactorCheckpointStorage(mongo, occurrentProperties.getSubscription().getCollection());
+    }
+
+    /**
+     * The zero-config {@link AppliedPositionStorage} a {@code @Projection(recordAppliedPosition = true)} resolves
+     * when the application declares none.
+     */
+    @Bean
+    @ConditionalOnMissingBean(AppliedPositionStorage.class)
+    public AppliedPositionStorage occurrentAppliedPositionStorage(ReactiveMongoOperations mongo, OccurrentProperties occurrentProperties) {
+        return new ReactiveMongoAppliedPositionStorage(mongo, occurrentProperties.getProjection().getAppliedPositionCollection());
     }
 
     /**